Implement the abstract class / derived class pattern for Storage
Currently the inheritdoc tag does not work in docfx, but users should be
referring to the StorageClient documentation anyway, and I definitely don't want
redundancy in manually-written documentation.
